+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/mixin/SpawnEggItemMixin.java
@@ -1,0 +1,32 @@
+package com.tristankechlo.random_mob_sizes.mixin;
+
+import com.tristankechlo.random_mob_sizes.mixin_helper.MobMixinAddon;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.SpawnEggItem;
+import net.minecraft.world.phys.Vec3;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+
+import java.util.Optional;
+
+@Mixin(SpawnEggItem.class)
+public abstract class SpawnEggItemMixin {
+
+    @Inject(method = "spawnOffspringFromSpawnEgg",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Mob;moveTo(DDDFF)V", shift = At.Shift.AFTER),
+            locals = LocalCapture.CAPTURE_FAILHARD)
+    private void spawnOffspringFromSpawnEgg$RandomMobSizes(
+            Player player, Mob parent, EntityType<? extends Mob> type, ServerLevel level, Vec3 pos, ItemStack stack,
+            CallbackInfoReturnable<Optional<Mob>> cir, Mob mob2) {
+        // apply scaling when child is created by right-clicking parent with spawn egg
+        ((MobMixinAddon) mob2).doFinalizeSpawn$RandomMobSizes(level);
+    }
+
+}