+++ Changelog.md
@@ -5,6 +5,7 @@
 - fix incorrect Piglin eye height
 - fix incorrect creeper model
 - fix cave spiders not having scaling
+- fix child mob scaling
 
 ### Version 1.18.2 - 2.0